pkg/payload: Move SummaryForReason to Error.Summary
There's no need for a separate function here, and using a method gives
us access to additional Error properties if we want them (although I
haven't pivoted to take advantage of that yet).
